requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pandas==2.1.3
openpyxl==3.1.2
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.53
langchain-openai==0.1.0
langsmith==0.1.17
lancedb==0.3.4
pyarrow==14.0.0
sentence-transformers==2.2.2
openai==1.12.0
pydantic==2.5.2
python-dotenv==1.0.0
huggingface-hub==0.19.4
requests==2.31.0